requirements.txt
pandas
python-dotenv
numpy 